Removed external git repositories from the docker image.
After consulting with Leah we arrived at the conclusion
that it is better to have copies of the repositories
installed locally, and have the workflow pull in the
necessary files. This in order to make sure that users
are able to use specific versions of these repositories
(which facilitates debugging). They will eventually be
incorporated into the pipeline.