Resolved validation warning
CWL changes the type of optional arguments to an array including
the type "null". When optional arguments are passed to CWL files
where these inputs are mandatory this can cause potential issues
and the runner warns against this. To resolve this the valueFrom
method should be used. See the CWL specification for details.